+++ app/src/db/migrations/0005_thankful_blackheart.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "status_pages" DROP COLUMN "branding_settings";